Automate npm releases on version bumps
