+++ .github/workflows/ci.yml
@@ -20,11 +20,6 @@ jobs:
           node-version: 22
           cache: npm
 
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.9"
-
       - name: Install dependencies
         run: npm ci
 
@@ -33,32 +28,3 @@ jobs:
 
       - name: Test
         run: npm test
-
-      - name: Test Python launcher
-        run: python -m unittest discover -s packages/openartifacts/test -p 'test_*.py'
-
-      - name: Verify npm distribution contents
-        run: node packages/openartifacts/test/verify-npm-dist.js
-
-      - name: Build Python distributions
-        run: |
-          python -m pip install build==1.3.0
-          python -m build --outdir dist/python packages/openartifacts
-
-      - name: Verify Python distribution contents
-        run: python packages/openartifacts/test/verify_python_dist.py dist/python
-
-  windows-python:
-    runs-on: windows-latest
-
-    steps:
-      - name: Check out repository
-        uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.9"
-
-      - name: Test Python launcher on Windows
-        run: python -m unittest discover -s packages/openartifacts/test -p 'test_*.py'
