+++ .github/workflows/release-openartifacts.yml
@@ -1,21 +1,74 @@
-name: Release OpenArtifacts packages
+name: Publish OpenArtifacts npm package
 
 on:
-  workflow_dispatch:
-    inputs:
-      version:
-        description: Stable X.Y.Z version in both package manifests
-        required: true
-        type: string
+  push:
+    branches:
+      - main
+    paths:
+      - packages/openartifacts/package.json
 
 concurrency:
-  group: release-openartifacts-packages
+  group: publish-openartifacts-npm
   cancel-in-progress: false
 
 permissions: {}
 
 jobs:
-  release:
+  check:
+    if: github.repository == 'Brevilabs/OpenArtifacts'
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+    outputs:
+      exists: ${{ steps.registry.outputs.exists }}
+      version: ${{ steps.package.outputs.version }}
+
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v4
+
+      - name: Set up Node.js
+        uses: actions/setup-node@v4
+        with:
+          node-version: 24
+
+      - name: Read and validate package version
+        id: package
+        run: |
+          set -euo pipefail
+          version="$(node -p "require('./packages/openartifacts/package.json').version")"
+          lock_version="$(node -p "require('./package-lock.json').packages['packages/openartifacts'].version")"
+          if [[ ! "$version" =~ ^(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)\.(0|[1-9][0-9]*)$ ]]; then
+            echo "::error::Package version must be a stable X.Y.Z version, found $version"
+            exit 1
+          fi
+          if [[ "$version" != "$lock_version" ]]; then
+            echo "::error::package.json is $version but package-lock.json is $lock_version"
+            exit 1
+          fi
+          echo "version=$version" >> "$GITHUB_OUTPUT"
+
+      - name: Check whether this version is already published
+        id: registry
+        env:
+          VERSION: ${{ steps.package.outputs.version }}
+        run: |
+          set -euo pipefail
+          code="$(curl --silent --show-error --location --retry 3 \
+            --output /tmp/npm-version.json --write-out '%{http_code}' \
+            "https://registry.npmjs.org/openartifacts/$VERSION")"
+          case "$code" in
+            200)
+              echo "exists=true" >> "$GITHUB_OUTPUT"
+              echo "openartifacts $VERSION is already published; nothing to do"
+              ;;
+            404) echo "exists=false" >> "$GITHUB_OUTPUT" ;;
+            *) echo "::error::npm registry returned HTTP $code"; exit 1 ;;
+          esac
+
+  publish:
+    needs: check
+    if: needs.check.outputs.exists != 'true'
     runs-on: ubuntu-latest
     environment: package-release
     permissions:
@@ -23,26 +76,8 @@ jobs:
       id-token: write
 
     steps:
-      - name: Refuse to release anything but main
-        if: github.ref != 'refs/heads/main'
-        env:
-          REF: ${{ github.ref }}
-        run: |
-          echo "::error::Refusing to release $REF. Packages release from main."
-          exit 1
-
       - name: Check out repository
         uses: actions/checkout@v4
-
-      - name: Set up Python
-        uses: actions/setup-python@v5
-        with:
-          python-version: "3.13"
-
-      - name: Validate requested version and package parity
-        env:
-          RELEASE_VERSION: ${{ inputs.version }}
-        run: python packages/openartifacts/scripts/check_version_parity.py --expect "$RELEASE_VERSION"
 
       - name: Set up Node.js
         uses: actions/setup-node@v4
@@ -56,75 +91,24 @@ jobs:
       - name: Install JavaScript dependencies
         run: npm ci
 
-      - name: Typecheck JavaScript
+      - name: Typecheck
         run: npm run typecheck
 
-      - name: Test JavaScript
+      - name: Test
         run: npm test
 
-      - name: Test Python launcher
-        run: python -m unittest discover -s packages/openartifacts/test -p 'test_*.py'
+      - name: Inspect package contents
+        run: npm pack --workspace packages/openartifacts --dry-run
 
-      - name: Verify npm distribution contents
-        run: node packages/openartifacts/test/verify-npm-dist.js
-
-      - name: Build Python distributions
-        run: |
-          python -m pip install build==1.3.0
-          python -m build --outdir dist/python packages/openartifacts
-
-      - name: Verify Python distribution contents
-        run: python packages/openartifacts/test/verify_python_dist.py dist/python
-
-      - name: Check registry versions
-        id: registry
-        env:
-          RELEASE_VERSION: ${{ inputs.version }}
-        run: |
-          set -euo pipefail
-
-          npm_code="$(curl --silent --show-error --location --retry 3 \
-            --output /tmp/npm-version.json --write-out '%{http_code}' \
-            "https://registry.npmjs.org/openartifacts/$RELEASE_VERSION")"
-          case "$npm_code" in
-            200) npm_exists=true ;;
-            404) npm_exists=false ;;
-            *) echo "::error::npm registry returned HTTP $npm_code"; exit 1 ;;
-          esac
-
-          pypi_code="$(curl --silent --show-error --location --retry 3 \
-            --output /tmp/pypi-version.json --write-out '%{http_code}' \
-            "https://pypi.org/pypi/openartifacts/$RELEASE_VERSION/json")"
-          case "$pypi_code" in
-            200) pypi_exists=true ;;
-            404) pypi_exists=false ;;
-            *) echo "::error::PyPI returned HTTP $pypi_code"; exit 1 ;;
-          esac
-
-          echo "npm_exists=$npm_exists" >> "$GITHUB_OUTPUT"
-          echo "pypi_exists=$pypi_exists" >> "$GITHUB_OUTPUT"
-
-      - name: Publish npm
-        if: steps.registry.outputs.npm_exists != 'true'
+      - name: Publish
         run: npm publish --workspace packages/openartifacts --access public --provenance
 
-      - name: Publish PyPI
-        if: steps.registry.outputs.pypi_exists != 'true'
-        uses: pypa/gh-action-pypi-publish@release/v1
-        with:
-          packages-dir: dist/python/
-
-      - name: Verify both registry versions
+      - name: Verify published version
         env:
-          RELEASE_VERSION: ${{ inputs.version }}
+          VERSION: ${{ needs.check.outputs.version }}
         run: |
-          set -euo pipefail
           curl --fail --silent --show-error --location --retry 10 \
             --retry-all-errors --retry-delay 3 \
-            "https://registry.npmjs.org/openartifacts/$RELEASE_VERSION" \
+            "https://registry.npmjs.org/openartifacts/$VERSION" \
             --output /dev/null
-          curl --fail --silent --show-error --location --retry 10 \
-            --retry-all-errors --retry-delay 3 \
-            "https://pypi.org/pypi/openartifacts/$RELEASE_VERSION/json" \
-            --output /dev/null
-          echo "openartifacts $RELEASE_VERSION is available from npm and PyPI"
+          echo "openartifacts $VERSION is available from npm"
