+++ docs/releasing.md
@@ -1,8 +1,8 @@
-# Releasing the official CLI packages
+# Releasing the official npm CLI
 
 This runbook is only for Brevilabs maintainers publishing the official
-`openartifacts` packages. Self-hosters do not need this workflow, a PyPI
-account, or Python; their Worker and npm CLI setup is unchanged.
+`openartifacts` npm package. Self-hosters do not need this workflow; their
+Worker and CLI setup is unchanged.
 
 ## One-time setup
 
@@ -10,9 +10,9 @@ Create a protected GitHub environment named `package-release` with:
 
 - required reviewer approval;
 - deployment branches restricted to `main`; and
-- no npm or PyPI tokens.
+- no npm token.
 
-Configure Trusted Publishing on both registries with the same identity:
+Configure npm Trusted Publishing with this identity:
 
 | Field | Value |
 | --- | --- |
@@ -21,39 +21,36 @@ Configure Trusted Publishing on both registries with the same identity:
 | Workflow | `release-openartifacts.yml` |
 | Environment | `package-release` |
 
-For the first PyPI release, configure a pending publisher for the unclaimed
-`openartifacts` project. npm Trusted Publishing is configured on the existing
-package; set its **Allowed action** to **npm publish**, because the staged-only
-default refuses a direct `npm publish`. Both registries authenticate the
-workflow through short-lived OIDC credentials, so GitHub stores no registry
-secret.
+Set its **Allowed action** to **npm publish**, because the staged-only default
+refuses a direct `npm publish`. npm authenticates the workflow through a
+short-lived OIDC credential, so GitHub stores no registry secret.
 
 ## Release
 
-1. From the repository root, update the npm manifest and lockfile together:
+From the repository root, update the manifest and lockfile together:
 
-   ```bash
-   npm version X.Y.Z --workspace packages/openartifacts --no-git-tag-version
-   ```
+```bash
+npm version X.Y.Z --workspace packages/openartifacts --no-git-tag-version
+```
 
-   Set `packages/openartifacts/pyproject.toml` to the same version.
-2. Merge that version to `main`.
-3. Run **Release OpenArtifacts packages** on `main` and enter that exact
-   version.
-4. Approve the single `package-release` environment prompt.
+Open and merge the version-bump PR. A change to
+`packages/openartifacts/package.json` on `main` automatically starts **Publish
+OpenArtifacts npm package**. Approve the `package-release` environment prompt;
+there is no separate workflow dispatch or version input.
 
-The workflow validates the ref and both manifests before contacting either
-registry. It runs the JavaScript and Python checks, builds both distributions,
-publishes npm first, waits until the exact version is visible, and then
-publishes PyPI. Registry versions are immutable. If a run stops between the two
-publishes, rerun the same version: an existing version is verified and skipped,
-so the missing registry can finish without a version bump.
+The workflow requires a stable `X.Y.Z` version, checks that the lockfile agrees,
+and looks up that version on npm. A new version runs the JavaScript checks,
+inspects the tarball, publishes with provenance, and verifies the registry. An
+already-published version exits successfully, so a rerun after publication is
+safe and a package-manifest edit without a version bump is a no-op.
 
-Prereleases are intentionally unsupported because npm SemVer and PyPI PEP 440
-do not preserve every prerelease string identically.
+PyPI is not part of this release process. The OpenArtifacts skill ships in the
+npm package and is also available directly to Hermes from its canonical
+repository path.
 
-The OpenArtifacts skill has no separate PyPI release. npm packages the
-canonical `packages/openartifacts/skill/openartifacts/SKILL.md`, while Hermes
-installs that same file directly from this repository through its native skill
-manager and uses `npx` when the CLI is not already installed. Maintainers do not
-copy or publish a second skill, and Hermes users do not need PyPI first.
+## One-time PyPI namespace package
+
+`pypi/openartifacts` is a small functional compatibility launcher for the
+official npm CLI. Its `0.0.1` version is intentionally independent from npm and
+is published manually once to establish the official PyPI namespace. Do not add
+it to the npm release workflow or keep its version in sync with npm.
