--- packages/openartifacts/pyproject.toml
@@ -1,40 +1,0 @@
-[build-system]
-requires = ["setuptools>=77"]
-build-backend = "setuptools.build_meta"
-
-[project]
-name = "openartifacts"
-version = "0.2.1"
-description = "Thin Python launcher for the OpenArtifacts npm CLI"
-readme = "python/README.md"
-requires-python = ">=3.9"
-license = "MIT"
-license-files = ["LICENSE"]
-authors = [
-  { name = "Brevilabs" },
-]
-keywords = ["agents", "artifacts", "cli", "markdown", "publishing"]
-classifiers = [
-  "Development Status :: 3 - Alpha",
-  "Environment :: Console",
-  "Operating System :: OS Independent",
-  "Programming Language :: Python :: 3",
-  "Programming Language :: Python :: 3 :: Only",
-  "Topic :: Internet :: WWW/HTTP",
-]
-dependencies = []
-
-[project.urls]
-Homepage = "https://github.com/Brevilabs/OpenArtifacts"
-Repository = "https://github.com/Brevilabs/OpenArtifacts.git"
-Issues = "https://github.com/Brevilabs/OpenArtifacts/issues"
-
-[project.scripts]
-openartifacts = "openartifacts_launcher.cli:entrypoint"
-
-[tool.setuptools]
-package-dir = { "" = "python" }
-
-[tool.setuptools.packages.find]
-where = ["python"]
-include = ["openartifacts_launcher"]
