--- packages/openartifacts/python/openartifacts_launcher/__init__.py
@@ -1,1 +1,0 @@
-"""Internal launcher for the OpenArtifacts npm CLI."""
