+++ pypi/openartifacts/pyproject.toml
@@ -1,0 +1,27 @@
+[build-system]
+requires = ["setuptools>=77"]
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "openartifacts"
+version = "0.0.1"
+description = "Official PyPI compatibility launcher for the OpenArtifacts npm CLI"
+readme = "README.md"
+requires-python = ">=3.9"
+license = "MIT"
+license-files = ["LICENSE"]
+authors = [
+  { name = "Brevilabs" },
+]
+keywords = ["agents", "artifacts", "cli", "publishing"]
+dependencies = []
+
+[project.urls]
+Homepage = "https://github.com/Brevilabs/OpenArtifacts"
+Repository = "https://github.com/Brevilabs/OpenArtifacts.git"
+
+[project.scripts]
+openartifacts = "openartifacts_pypi:entrypoint"
+
+[tool.setuptools]
+py-modules = ["openartifacts_pypi"]
